more bugfixes
git-svn-id: https://svn.kwarc.info/repos/MMT@2702 09069c07-60fa-4895-8316-727242e252e4

+++ src/mmt-api/trunk/src/main/info/kwarc/mmt/api/uom/StandardLiterals.scala
@@ -67,7 +67,7 @@ object Quotient {
 /** bundles functions that are typically used when defining literals based on integers */
 trait IntegerLiteral extends SemanticType {
    type univ = BigInt
-   def fromString(s: String) = s.toInt
+   def fromString(s: String) = BigInt(s)
    override def lex = Some(new parser.NumberLiteralLexer(false))
 }
